+++ src/main/java/org/riekr/jloga/search/PlainTextComponent.java
@@ -1,5 +1,8 @@
 package org.riekr.jloga.search;
 
+import static org.riekr.jloga.react.Observer.uniq;
+import static org.riekr.jloga.utils.UIUtils.newToggleButton;
+
 import javax.swing.*;
 import java.util.function.Consumer;
 
@@ -9,7 +12,6 @@ import org.riekr.jloga.misc.SearchComboEntry;
 import org.riekr.jloga.react.Unsubscribable;
 import org.riekr.jloga.search.simple.SimpleSearchPredicate;
 import org.riekr.jloga.ui.MRUTextCombo;
-import org.riekr.jloga.utils.UIUtils;
 
 public class PlainTextComponent extends Box implements SearchComponent {
 	private static final long serialVersionUID = -2002183911884676582L;
@@ -24,18 +26,18 @@ public class PlainTextComponent extends Box implements SearchComponent {
 		_combo = new MRUTextCombo<>("plainTextSearch." + level, SearchComboEntry::new);
 		add(_combo);
 		SearchComboEntry initialValue = _combo.getValue();
-		JToggleButton negateBtn = UIUtils.newToggleButton("!", "Negate", initialValue.negate, (b) -> {
+		JToggleButton negateBtn = newToggleButton("!", "Negate", initialValue.negate, (b) -> {
 			_combo.getValue().negate = b;
 			_combo.save();
 		});
-		JToggleButton caseBtn = UIUtils.newToggleButton("\uD83D\uDDDA", "Case insensitive", initialValue.caseInsensitive, (b) -> {
+		JToggleButton caseBtn = newToggleButton("\uD83D\uDDDA", "Case insensitive", initialValue.caseInsensitive, (b) -> {
 			_combo.getValue().caseInsensitive = b;
 			_combo.save();
 		});
-		_combo.subject.subscribe((value) -> {
+		_combo.subject.subscribe(uniq((value) -> {
 			negateBtn.setSelected(value.negate);
 			caseBtn.setSelected(value.caseInsensitive);
-		});
+		}));
 		add(negateBtn);
 		add(caseBtn);
 	}